requirements.txt
# collect.py - 구글 뉴스 RSS 파싱
feedparser

# collect.py 금감원 웹 크롤링 http 요청
requests

# collect.py html 파싱
beautifulsoup4

# beautifulsoup4 파서 (html.parser 대체, 속도향상)
lxml

# summarize.py - claude API 클라이언트
groq

# flask - kakao.py - 카카오 스킬 서버
flask

# gunicorn - Render 프로덕션 서버
gunicorn

# .env 파일로 key 관리
python-dotenv
